+++ chef/resource_data_bag.go
@@ -23,6 +23,9 @@ func resourceChefDataBag() *schema.Resource {
 				Computed: true,
 			},
 		},
+		Importer: &schema.ResourceImporter{
+			State: schema.ImportStatePassthrough,
+		},
 	}
 }