+++ chef/resource_environment.go
@@ -47,6 +47,9 @@ func resourceChefEnvironment() *schema.Resource {
 				},
 			},
 		},
+		Importer: &schema.ResourceImporter{
+			State: schema.ImportStatePassthrough,
+		},
 	}
 }