+++ chef/resource_role.go
@@ -48,6 +48,9 @@ func resourceChefRole() *schema.Resource {
 				},
 			},
 		},
+		Importer: &schema.ResourceImporter{
+			State: schema.ImportStatePassthrough,
+		},
 	}
 }